[FIX] bact bessyii ophyd: separate installable package
